+++ exo-go/features/template/fetch_templates/fetch_templates.feature
@@ -4,7 +4,7 @@ Feature: fetching updates for existing templates
 	I want to be able to fetch the latest updates for the existing service templates in application codebase
 	So that I don't have to manually run the git command to update the submodules
 
-  - run "exo fetch-templates" to fetch the latest updates for the existing service templates
+  - run "exo template fetch" to fetch the latest updates for the existing service templates
 
   Scenario: fetching updates for all existing service templates
     Given I am in the root directory of an empty application called "test app"
@@ -12,5 +12,5 @@ Feature: fetching updates for existing templates
     And my application has the templates:
       | NAME        | URL                                      |
       | boilr-spark | https://github.com/tmrts/boilr-spark.git |
-    When running "exo fetch-templates" in my application directory
+    When running "exo template fetch" in my application directory
     Then it prints "done" in the terminal
