+++ exo-go/features/template/remove_template/remove_template.feature
@@ -4,7 +4,7 @@ Feature: removing existing templates
 	I want to be able to remove existing service templates from my application codebase
 	So that I don't have to manually run the git command to remove the submodules
 
-  - run "exo remove-template <name>" to remove an existing service template from the application codebase
+  - run "exo template remove <name>" to remove an existing service template from the application codebase
 
 
   Background:
@@ -16,12 +16,12 @@ Feature: removing existing templates
     Given my application has the templates:
       | NAME        | URL                                      |
       | boilr-spark | https://github.com/tmrts/boilr-spark.git |
-    When running "exo remove-template boilr-spark" in my application directory
+    When running "exo template remove boilr-spark" in my application directory
     Then my git repository does not have any submodules
 
 
   Scenario: removing a non-existing service template
-    When starting "exo remove-template non-existing" in my application directory
+    When starting "exo template remove non-existing" in my application directory
     Then I eventually see:
       """
       Error: template does not exist
